Separate out PgOutput again and name appropriately
