+++ src/Npgsql/Replication/PgOutput/Messages/DeleteMessage.cs
@@ -1,7 +1,7 @@
-﻿using NpgsqlTypes;
-using System;
+﻿using System;
+using NpgsqlTypes;
 
-namespace Npgsql.Replication.Messages
+namespace Npgsql.Replication.PgOutput.Messages
 {
     /// <summary>
     /// Abstract base class for Logical Replication Protocol delete message types.
