+++ src/Npgsql/Replication/PgOutput/Messages/RelationMessageColumn.cs
@@ -1,4 +1,4 @@
-﻿namespace Npgsql.Replication.Messages
+﻿namespace Npgsql.Replication.PgOutput.Messages
 {
     /// <summary>
     /// Represents a column in a Logical Replication Protocol relation message
