+++ src/Npgsql/Replication/PgOutput/NpgsqlPgOutputReplicationSlot.cs
@@ -1,45 +1,45 @@
 ﻿using Npgsql.Replication.Internal;
 
-namespace Npgsql.Replication
+namespace Npgsql.Replication.PgOutput
 {
     /// <summary>
     /// Acts as a proxy for a logical replication slot initialized for for the logical streaming replication protocol
     /// (pgoutput logical decoding plugin).
     /// </summary>
-    public class NpgsqlLogicalReplicationSlot : NpgsqlLogicalReplicationSlotBase
+    public class NpgsqlPgOutputReplicationSlot : NpgsqlLogicalReplicationSlot
     {
         /// <summary>
-        /// Creates a new <see cref="NpgsqlLogicalReplicationSlot"/> instance.
+        /// Creates a new <see cref="NpgsqlPgOutputReplicationSlot"/> instance.
         /// </summary>
         /// <remarks>
-        /// Create a <see cref="NpgsqlLogicalReplicationSlot"/> instance with this
+        /// Create a <see cref="NpgsqlPgOutputReplicationSlot"/> instance with this
         /// constructor to wrap an existing PostgreSQL replication slot that has
         /// been initialized for the pgoutput logical decoding plugin.
         /// </remarks>
         /// <param name="slotName">The name of the existing replication slot</param>
-        public NpgsqlLogicalReplicationSlot(string slotName)
+        public NpgsqlPgOutputReplicationSlot(string slotName)
             : this(new NpgsqlReplicationSlotOptions(slotName)) { }
 
         /// <summary>
-        /// Creates a new <see cref="NpgsqlLogicalReplicationSlot"/> instance.
+        /// Creates a new <see cref="NpgsqlPgOutputReplicationSlot"/> instance.
         /// </summary>
         /// <remarks>
-        /// Create a <see cref="NpgsqlLogicalReplicationSlot"/> instance with this
+        /// Create a <see cref="NpgsqlPgOutputReplicationSlot"/> instance with this
         /// constructor to wrap an existing PostgreSQL replication slot that has
         /// been initialized for the pgoutput logical decoding plugin.
         /// </remarks>
         /// <param name="options">The <see cref="NpgsqlReplicationSlotOptions"/> representing the existing replication slot</param>
-        public NpgsqlLogicalReplicationSlot(NpgsqlReplicationSlotOptions options) : base("pgoutput", options) { }
+        public NpgsqlPgOutputReplicationSlot(NpgsqlReplicationSlotOptions options) : base("pgoutput", options) { }
 
         /// <summary>
-        /// Creates a new <see cref="NpgsqlLogicalReplicationSlot"/> instance.
+        /// Creates a new <see cref="NpgsqlPgOutputReplicationSlot"/> instance.
         /// </summary>
         /// <remarks>
         /// This constructor is intended to be consumed by plugins sitting on top of
-        /// <see cref="NpgsqlLogicalReplicationSlot"/>
+        /// <see cref="NpgsqlPgOutputReplicationSlot"/>
         /// </remarks>
-        /// <param name="slot">The <see cref="NpgsqlLogicalReplicationSlot"/> from which the new instance should be initialized</param>
-        protected NpgsqlLogicalReplicationSlot(NpgsqlLogicalReplicationSlot slot)
+        /// <param name="slot">The <see cref="NpgsqlPgOutputReplicationSlot"/> from which the new instance should be initialized</param>
+        protected NpgsqlPgOutputReplicationSlot(NpgsqlPgOutputReplicationSlot slot)
             : base(slot.OutputPlugin, new NpgsqlReplicationSlotOptions(slot.SlotName, slot.ConsistentPoint, slot.SnapshotName)) { }
     }
 }
